server/requirements.txt
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.64
botocore==1.40.64
Cartopy==0.24.1
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
eventregistry==9.1
executing==2.1.0
fastjsonschema==2.20.0
finance-utils==1.0.3.dev5
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.54.1
frozendict==2.4.6
idna==3.11
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
multitasking==0.0.12
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.4
packaging==24.1
pandas==2.3.3
panndas==0.0.2
parso==0.8.4
peewee==3.18.2
pendulum==3.1.0
pillow==10.4.0
platformdirs==4.3.2
plotly==5.24.1
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.47
protobuf==6.33.0
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.23
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.4
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.1
pywin32==306
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.5
rpds-py==0.21.0
s3transfer==0.14.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.7
six==1.16.0
soupsieve==2.8
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.66